% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{BatchPredictRequest.params}
\alias{BatchPredictRequest.params}
\title{BatchPredictRequest.params Object}
\usage{
BatchPredictRequest.params()
}
\value{
BatchPredictRequest.params object
}
\description{
BatchPredictRequest.params Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific parameters for the predictions, any string mustbe up to 25000 characters long.*  For Image Classification:   `score_threshold` - (float) A value from 0.0 to 1.0. When the model        makes predictions for an image, it will only produce results that        have at least this confidence score. The default is 0.5.*  For Image Object Detection:   `score_threshold` - (float) When Model detects objects on the image,       it will only produce bounding boxes which have at least this       confidence score. Value in 0 to 1 range, default is 0.5.   `max_bounding_box_count` - (int64) No more than this number of bounding       boxes will be produced per image. Default is 100, the       requested value may be limited by server.*  For Video Classification :   `score_threshold` - (float) A value from 0.0 to 1.0. When the model       makes predictions for a video, it will only produce results that       have at least this confidence score. The default is 0.5.   `segment_classification` - (boolean) Set to true to request       segment-level classification. AutoML Video Intelligence returns       labels and their confidence scores for the entire segment of the       video that user specified in the request configuration.       The default is 'true'.   `shot_classification` - (boolean) Set to true to request shot-level       classification. AutoML Video Intelligence determines the boundaries       for each camera shot in the entire segment of the video that user       specified in the request configuration. AutoML Video Intelligence       then returns labels and their confidence scores for each detected       shot, along with the start and end time of the shot.       WARNING: Model evaluation is not done for this classification type,       the quality of it depends on training data, but there are no metrics       provided to describe that quality. The default is 'false'.   `1s_interval_classification` - (boolean) Set to true to request       classification for a video at one-second intervals. AutoML Video       Intelligence returns labels and their confidence scores for each       second of the entire segment of the video that user specified in the       request configuration.       WARNING: Model evaluation is not done for this classification       type, the quality of it depends on training data, but there are no       metrics provided to describe that quality. The default is       'false'.*  For Video Object Tracking:   `score_threshold` - (float) When Model detects objects on video frames,       it will only produce bounding boxes which have at least this       confidence score. Value in 0 to 1 range, default is 0.5.   `max_bounding_box_count` - (int64) No more than this number of bounding       boxes will be returned per frame. Default is 100, the requested       value may be limited by server.   `min_bounding_box_size` - (float) Only bounding boxes with shortest edge     at least that long as a relative value of video frame size will be     returned. Value in 0 to 1 range. Default is 0.
}
\seealso{
Other BatchPredictRequest functions: \code{\link{BatchPredictRequest}},
  \code{\link{projects.locations.models.batchPredict}}
}
\concept{BatchPredictRequest functions}
