% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{CategoryStats}
\alias{CategoryStats}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
CategoryStats(topCategoryStats = NULL)
}
\arguments{
\item{topCategoryStats}{The statistics of the top 20 CATEGORY values, ordered by}
}
\value{
CategoryStats object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2020-03-30 21:03:29
filename: /Users/justin/dev/R/autoGoogleAPI/googleautomlv1beta1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
CategoryStats Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data statistics of a series of CATEGORY values.
}
\concept{CategoryStats functions}
