% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ClassificationEvaluationMetrics}
\alias{ClassificationEvaluationMetrics}
\title{ClassificationEvaluationMetrics Object}
\usage{
ClassificationEvaluationMetrics(auRoc = NULL, baseAuPrc = NULL,
  annotationSpecId = NULL, logLoss = NULL, auPrc = NULL,
  confusionMatrix = NULL, confidenceMetricsEntry = NULL)
}
\arguments{
\item{auRoc}{Output only}

\item{baseAuPrc}{Output only}

\item{annotationSpecId}{Output only}

\item{logLoss}{Output only}

\item{auPrc}{Output only}

\item{confusionMatrix}{Output only}

\item{confidenceMetricsEntry}{Output only}
}
\value{
ClassificationEvaluationMetrics object
}
\description{
ClassificationEvaluationMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model evaluation metrics for classification problems.Note: For Video Classification this metrics only describe quality of theVideo Classification predictions of 'segment_classification' type.
}
\concept{ClassificationEvaluationMetrics functions}
