% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DataType}
\alias{DataType}
\title{DataType Object}
\usage{
DataType(
  typeCode = NULL,
  structType = NULL,
  timeFormat = NULL,
  nullable = NULL,
  listElementType = NULL
)
}
\arguments{
\item{typeCode}{Required}

\item{structType}{If type_code == STRUCT, then `struct_type` provides type information for the struct's fields}

\item{timeFormat}{If type_code == TIMESTAMP then `time_format` provides the format in which that time field is expressed}

\item{nullable}{If true, this DataType can also be `NULL`}

\item{listElementType}{If type_code == ARRAY, then `list_element_type` is the type of the elements}
}
\value{
DataType object
}
\description{
DataType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Indicated the type of data that can be stored in a structured data entity (e.g. a table).
}
\concept{DataType functions}
