% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Document}
\alias{Document}
\title{Document Object}
\usage{
Document(documentText = NULL, layout = NULL, pageCount = NULL,
  documentDimensions = NULL, inputConfig = NULL)
}
\arguments{
\item{documentText}{The plain text version of this document}

\item{layout}{Describes the layout of the document}

\item{pageCount}{Number of pages in the document}

\item{documentDimensions}{The dimensions of the page in the document}

\item{inputConfig}{An input config specifying the content of the document}
}
\value{
Document object
}
\description{
Document Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A structured text document e.g. a PDF.
}
\concept{Document functions}
