% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{DocumentDimensions}
\alias{DocumentDimensions}
\title{DocumentDimensions Object}
\usage{
DocumentDimensions(width = NULL, height = NULL, unit = NULL)
}
\arguments{
\item{width}{Width value of the document, works together with the unit}

\item{height}{Height value of the document, works together with the unit}

\item{unit}{Unit of the dimension}
}
\value{
DocumentDimensions object
}
\description{
DocumentDimensions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message that describes dimension of a document.
}
\concept{DocumentDimensions functions}
