% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ExportDataOperationMetadata}
\alias{ExportDataOperationMetadata}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
ExportDataOperationMetadata(outputInfo = NULL)
}
\arguments{
\item{outputInfo}{Output only}
}
\value{
ExportDataOperationMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-06-27 22:31:58
filename: /Users/justin/Dropbox/dev/R/autoGoogleAPI/googleautomlv1beta1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ExportDataOperationMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Details of ExportData operation.
}
\concept{ExportDataOperationMetadata functions}
