% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{GcsSource}
\alias{GcsSource}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
GcsSource(inputUris = NULL)
}
\arguments{
\item{inputUris}{Required}
}
\value{
GcsSource object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-06-22 17:31:36
filename: /Users/justin/Dropbox/dev/R/autoGoogleAPI/googleautomlv1beta1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GcsSource Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Google Cloud Storage location for the input content.
}
\concept{GcsSource functions}
