% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{HistogramBucket}
\alias{HistogramBucket}
\title{HistogramBucket Object}
\usage{
HistogramBucket(min = NULL, count = NULL, max = NULL)
}
\arguments{
\item{min}{The minimum value of the bucket, inclusive}

\item{count}{The number of data values that are in the bucket, i}

\item{max}{The maximum value of the bucket, exclusive unless max = `'Infinity'`, in which case it's inclusive}
}
\value{
HistogramBucket object
}
\description{
HistogramBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A bucket of a histogram.
}
\concept{HistogramBucket functions}
