% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelMetadata}
\alias{ImageClassificationModelMetadata}
\title{ImageClassificationModelMetadata Object}
\usage{
ImageClassificationModelMetadata(modelType = NULL, trainBudget = NULL,
  trainCost = NULL, stopReason = NULL, baseModelId = NULL)
}
\arguments{
\item{modelType}{Optional}

\item{trainBudget}{Required}

\item{trainCost}{Output only}

\item{stopReason}{Output only}

\item{baseModelId}{Optional}
}
\value{
ImageClassificationModelMetadata object
}
\description{
ImageClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata for image classification.
}
\concept{ImageClassificationModelMetadata functions}
