% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Layout}
\alias{Layout}
\title{Layout Object}
\usage{
Layout(pageNumber = NULL, boundingPoly = NULL,
  textSegmentType = NULL, textSegment = NULL)
}
\arguments{
\item{pageNumber}{Page number of the text_segment in the original document, starts}

\item{boundingPoly}{The position of the text_segment in the page}

\item{textSegmentType}{The type of the text_segment in document}

\item{textSegment}{Text Segment that represents a segment in}
}
\value{
Layout object
}
\description{
Layout Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the layout information of a text_segment in the document.
}
\concept{Layout functions}
