% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ListTableSpecsResponse}
\alias{ListTableSpecsResponse}
\title{ListTableSpecsResponse Object}
\usage{
ListTableSpecsResponse(tableSpecs = NULL, nextPageToken = NULL)
}
\arguments{
\item{tableSpecs}{The table specs read}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListTableSpecsResponse object
}
\description{
ListTableSpecsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for AutoMl.ListTableSpecs.
}
\concept{ListTableSpecsResponse functions}
