% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Model}
\alias{Model}
\title{Model Object}
\usage{
Model(name = NULL, translationModelMetadata = NULL,
  videoClassificationModelMetadata = NULL, displayName = NULL,
  deploymentState = NULL, imageClassificationModelMetadata = NULL,
  createTime = NULL, updateTime = NULL,
  textExtractionModelMetadata = NULL, tablesModelMetadata = NULL,
  textClassificationModelMetadata = NULL, datasetId = NULL,
  imageObjectDetectionModelMetadata = NULL,
  textSentimentModelMetadata = NULL,
  videoObjectTrackingModelMetadata = NULL)
}
\arguments{
\item{name}{Output only}

\item{translationModelMetadata}{Metadata for translation models}

\item{videoClassificationModelMetadata}{Metadata for video classification models}

\item{displayName}{Required}

\item{deploymentState}{Output only}

\item{imageClassificationModelMetadata}{Metadata for image classification models}

\item{createTime}{Output only}

\item{updateTime}{Output only}

\item{textExtractionModelMetadata}{Metadata for text extraction models}

\item{tablesModelMetadata}{Metadata for Tables models}

\item{textClassificationModelMetadata}{Metadata for text classification models}

\item{datasetId}{Required}

\item{imageObjectDetectionModelMetadata}{Metadata for image object detection models}

\item{textSentimentModelMetadata}{Metadata for text sentiment models}

\item{videoObjectTrackingModelMetadata}{Metadata for video object tracking models}
}
\value{
Model object
}
\description{
Model Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
API proto representing a trained machine learning model.
}
\seealso{
Other Model functions: \code{\link{projects.locations.models.create}}
}
\concept{Model functions}
