% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(
  undeployModelDetails = NULL,
  createTime = NULL,
  createModelDetails = NULL,
  updateTime = NULL,
  deleteDetails = NULL,
  importDataDetails = NULL,
  progressPercent = NULL,
  exportModelDetails = NULL,
  exportDataDetails = NULL,
  batchPredictDetails = NULL,
  deployModelDetails = NULL,
  partialFailures = NULL,
  exportEvaluatedExamplesDetails = NULL
)
}
\arguments{
\item{undeployModelDetails}{Details of an UndeployModel operation}

\item{createTime}{Output only}

\item{createModelDetails}{Details of CreateModel operation}

\item{updateTime}{Output only}

\item{deleteDetails}{Details of a Delete operation}

\item{importDataDetails}{Details of ImportData operation}

\item{progressPercent}{Output only}

\item{exportModelDetails}{Details of ExportModel operation}

\item{exportDataDetails}{Details of ExportData operation}

\item{batchPredictDetails}{Details of BatchPredict operation}

\item{deployModelDetails}{Details of a DeployModel operation}

\item{partialFailures}{Output only}

\item{exportEvaluatedExamplesDetails}{Details of ExportEvaluatedExamples operation}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata used across all long running operations returned by AutoML API.
}
\concept{OperationMetadata functions}
