% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictResponse.metadata}
\alias{PredictResponse.metadata}
\title{PredictResponse.metadata Object}
\usage{
PredictResponse.metadata()
}
\value{
PredictResponse.metadata object
}
\description{
PredictResponse.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional domain-specific prediction response metadata. * For Image Object Detection: `max_bounding_box_count` - (int64) At most that many bounding boxes per image could have been returned. * For Text Sentiment: `sentiment_score` - (float, deprecated) A value between -1 and 1, -1 maps to least positive sentiment, while 1 maps to the most positive one and the higher the score, the more positive the sentiment in the document is. Yet these values are relative to the training data, so e.g. if all data was positive then -1 will be also positive (though the least). The sentiment_score shouldn't be confused with 'score' or 'magnitude' from the previous Natural Language Sentiment Analysis API.
}
\seealso{
Other PredictResponse functions: 
\code{\link{PredictResponse}()}
}
\concept{PredictResponse functions}
