% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{PredictResponse}
\alias{PredictResponse}
\title{PredictResponse Object}
\usage{
PredictResponse(PredictResponse.metadata = NULL, metadata = NULL,
  payload = NULL)
}
\arguments{
\item{PredictResponse.metadata}{The \link{PredictResponse.metadata} object or list of objects}

\item{metadata}{Additional domain-specific prediction response metadata}

\item{payload}{Prediction result}
}
\value{
PredictResponse object
}
\description{
PredictResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for PredictionService.Predict.
}
\seealso{
Other PredictResponse functions: \code{\link{PredictResponse.metadata}}
}
\concept{PredictResponse functions}
