% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Row}
\alias{Row}
\title{Row Object}
\usage{
Row(columnSpecIds = NULL, values = NULL)
}
\arguments{
\item{columnSpecIds}{Input Only}

\item{values}{Input Only}
}
\value{
Row object
}
\description{
Row Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A representation of a row in a relational table.
}
\concept{Row functions}
