% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{Status}
\alias{Status}
\title{Status Object}
\usage{
Status(message = NULL, details = NULL, code = NULL)
}
\arguments{
\item{message}{A developer-facing error message, which should be in English}

\item{details}{A list of messages that carry the error details}

\item{code}{The status code, which should be an enum value of google}
}
\value{
Status object
}
\description{
Status Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).
}
\concept{Status functions}
