% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesAnnotation}
\alias{TablesAnnotation}
\title{TablesAnnotation Object}
\usage{
TablesAnnotation(
  value = NULL,
  baselineScore = NULL,
  tablesModelColumnInfo = NULL,
  predictionInterval = NULL,
  score = NULL
)
}
\arguments{
\item{value}{The predicted value of the row's target_column}

\item{baselineScore}{Output only}

\item{tablesModelColumnInfo}{Output only}

\item{predictionInterval}{Output only}

\item{score}{Output only}
}
\value{
TablesAnnotation object
}
\description{
TablesAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains annotation details specific to Tables.
}
\concept{TablesAnnotation functions}
