% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesModelMetadata}
\alias{TablesModelMetadata}
\title{TablesModelMetadata Object}
\usage{
TablesModelMetadata(
  optimizationObjectiveRecallValue = NULL,
  inputFeatureColumnSpecs = NULL,
  disableEarlyStopping = NULL,
  targetColumnSpec = NULL,
  tablesModelColumnInfo = NULL,
  trainBudgetMilliNodeHours = NULL,
  trainCostMilliNodeHours = NULL,
  optimizationObjectivePrecisionValue = NULL,
  optimizationObjective = NULL
)
}
\arguments{
\item{optimizationObjectiveRecallValue}{Required when optimization_objective is 'MAXIMIZE_PRECISION_AT_RECALL'}

\item{inputFeatureColumnSpecs}{Column specs of the dataset's primary table's columns, on which the model is trained and which are used as the input for predictions}

\item{disableEarlyStopping}{Use the entire training budget}

\item{targetColumnSpec}{Column spec of the dataset's primary table's column the model is predicting}

\item{tablesModelColumnInfo}{Output only}

\item{trainBudgetMilliNodeHours}{Required}

\item{trainCostMilliNodeHours}{Output only}

\item{optimizationObjectivePrecisionValue}{Required when optimization_objective is 'MAXIMIZE_RECALL_AT_PRECISION'}

\item{optimizationObjective}{Objective function the model is optimizing towards}
}
\value{
TablesModelMetadata object
}
\description{
TablesModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Tables.
}
\concept{TablesModelMetadata functions}
