% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TableSpec}
\alias{TableSpec}
\title{TableSpec Object}
\usage{
TableSpec(
  inputConfigs = NULL,
  rowCount = NULL,
  columnCount = NULL,
  etag = NULL,
  validRowCount = NULL,
  timeColumnSpecId = NULL,
  name = NULL
)
}
\arguments{
\item{inputConfigs}{Output only}

\item{rowCount}{Output only}

\item{columnCount}{Output only}

\item{etag}{Used to perform consistent read-modify-write updates}

\item{validRowCount}{Output only}

\item{timeColumnSpecId}{column_spec_id of the time column}

\item{name}{Output only}
}
\value{
TableSpec object
}
\description{
TableSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A specification of a relational table. The table's schema is represented via its child column specs. It is pre-populated as part of ImportData by schema inference algorithm, the version of which is a required parameter of ImportData InputConfig. Note: While working with a table, at times the schema may be inconsistent with the data in the table (e.g. string in a FLOAT64 column). The consistency validation is done upon creation of a model. Used by: * Tables
}
\seealso{
Other TableSpec functions: 
\code{\link{projects.locations.datasets.tableSpecs.patch}()}
}
\concept{TableSpec functions}
