% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionEvaluationMetricsConfidenceMetricsEntry}
\alias{TextExtractionEvaluationMetricsConfidenceMetricsEntry}
\title{TextExtractionEvaluationMetricsConfidenceMetricsEntry Object}
\usage{

  TextExtractionEvaluationMetricsConfidenceMetricsEntry(confidenceThreshold = NULL,
  recall = NULL, f1Score = NULL, precision = NULL)
}
\arguments{
\item{confidenceThreshold}{Output only}

\item{recall}{Output only}

\item{f1Score}{Output only}

\item{precision}{Output only}
}
\value{
TextExtractionEvaluationMetricsConfidenceMetricsEntry object
}
\description{
TextExtractionEvaluationMetricsConfidenceMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for a single confidence threshold.
}
\concept{TextExtractionEvaluationMetricsConfidenceMetricsEntry functions}
