% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextExtractionModelMetadata}
\alias{TextExtractionModelMetadata}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
TextExtractionModelMetadata()
}
\value{
TextExtractionModelMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-07-03 22:24:43
filename: /Users/justin/Dropbox/dev/R/autoGoogleAPI/googleautomlv1beta1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
TextExtractionModelMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata that is specific to text extraction.
}
\concept{TextExtractionModelMetadata functions}
