% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TextSegment}
\alias{TextSegment}
\title{TextSegment Object}
\usage{
TextSegment(content = NULL, endOffset = NULL, startOffset = NULL)
}
\arguments{
\item{content}{Output only}

\item{endOffset}{Required}

\item{startOffset}{Required}
}
\value{
TextSegment object
}
\description{
TextSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A contiguous part of a text (string), assuming it has an UTF-8 NFC encoding.
}
\concept{TextSegment functions}
