% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TimeSegment}
\alias{TimeSegment}
\title{TimeSegment Object}
\usage{
TimeSegment(startTimeOffset = NULL, endTimeOffset = NULL)
}
\arguments{
\item{startTimeOffset}{Start of the time segment (inclusive), represented as the duration since the example start}

\item{endTimeOffset}{End of the time segment (exclusive), represented as the duration since the example start}
}
\value{
TimeSegment object
}
\description{
TimeSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A time period inside of an example that has a time dimension (e.g. video).
}
\concept{TimeSegment functions}
