% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TranslationModelMetadata}
\alias{TranslationModelMetadata}
\title{TranslationModelMetadata Object}
\usage{
TranslationModelMetadata(
  targetLanguageCode = NULL,
  baseModel = NULL,
  sourceLanguageCode = NULL
)
}
\arguments{
\item{targetLanguageCode}{Output only}

\item{baseModel}{The resource name of the model to use as a baseline to train the custom model}

\item{sourceLanguageCode}{Output only}
}
\value{
TranslationModelMetadata object
}
\description{
TranslationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata that is specific to translation.
}
\concept{TranslationModelMetadata functions}
