% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoObjectTrackingAnnotation}
\alias{VideoObjectTrackingAnnotation}
\title{VideoObjectTrackingAnnotation Object}
\usage{
VideoObjectTrackingAnnotation(
  timeOffset = NULL,
  instanceId = NULL,
  score = NULL,
  boundingBox = NULL
)
}
\arguments{
\item{timeOffset}{Required}

\item{instanceId}{Optional}

\item{score}{Output only}

\item{boundingBox}{Required}
}
\value{
VideoObjectTrackingAnnotation object
}
\description{
VideoObjectTrackingAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details for video object tracking.
}
\concept{VideoObjectTrackingAnnotation functions}
