% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoObjectTrackingDatasetMetadata}
\alias{VideoObjectTrackingDatasetMetadata}
\title{VideoObjectTrackingDatasetMetadata Object}
\usage{
VideoObjectTrackingDatasetMetadata()
}
\value{
VideoObjectTrackingDatasetMetadata object
}
\description{
VideoObjectTrackingDatasetMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata specific to video object tracking.
}
\concept{VideoObjectTrackingDatasetMetadata functions}
