% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoObjectTrackingModelMetadata}
\alias{VideoObjectTrackingModelMetadata}
\title{VideoObjectTrackingModelMetadata Object}
\usage{
VideoObjectTrackingModelMetadata()
}
\value{
VideoObjectTrackingModelMetadata object
}
\description{
VideoObjectTrackingModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to video object tracking.
}
\concept{VideoObjectTrackingModelMetadata functions}
