% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{AnnotationPayload}
\alias{AnnotationPayload}
\title{AnnotationPayload Object}
\usage{
AnnotationPayload(videoObjectTracking = NULL, tables = NULL,
  videoClassification = NULL, annotationSpecId = NULL,
  classification = NULL, textExtraction = NULL,
  imageObjectDetection = NULL, displayName = NULL,
  textSentiment = NULL, translation = NULL)
}
\arguments{
\item{videoObjectTracking}{Annotation details for video object tracking}

\item{tables}{Annotation details for Tables}

\item{videoClassification}{Annotation details for video classification}

\item{annotationSpecId}{Output only}

\item{classification}{Annotation details for content or image classification}

\item{textExtraction}{Annotation details for text extraction}

\item{imageObjectDetection}{Annotation details for image object detection}

\item{displayName}{Output only}

\item{textSentiment}{Annotation details for text sentiment}

\item{translation}{Annotation details for translation}
}
\value{
AnnotationPayload object
}
\description{
AnnotationPayload Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains annotation information that is relevant to AutoML.
}
\concept{AnnotationPayload functions}
