% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ClassificationEvaluationMetricsConfidenceMetricsEntry}
\alias{ClassificationEvaluationMetricsConfidenceMetricsEntry}
\title{ClassificationEvaluationMetricsConfidenceMetricsEntry Object}
\usage{
ClassificationEvaluationMetricsConfidenceMetricsEntry(
  precisionAt1 = NULL,
  truePositiveCount = NULL,
  f1ScoreAt1 = NULL,
  precision = NULL,
  falsePositiveCount = NULL,
  f1Score = NULL,
  trueNegativeCount = NULL,
  positionThreshold = NULL,
  falsePositiveRate = NULL,
  recall = NULL,
  confidenceThreshold = NULL,
  falsePositiveRateAt1 = NULL,
  falseNegativeCount = NULL,
  recallAt1 = NULL
)
}
\arguments{
\item{precisionAt1}{Output only}

\item{truePositiveCount}{Output only}

\item{f1ScoreAt1}{Output only}

\item{precision}{Output only}

\item{falsePositiveCount}{Output only}

\item{f1Score}{Output only}

\item{trueNegativeCount}{Output only}

\item{positionThreshold}{Output only}

\item{falsePositiveRate}{Output only}

\item{recall}{Output only}

\item{confidenceThreshold}{Output only}

\item{falsePositiveRateAt1}{Output only}

\item{falseNegativeCount}{Output only}

\item{recallAt1}{Output only}
}
\value{
ClassificationEvaluationMetricsConfidenceMetricsEntry object
}
\description{
ClassificationEvaluationMetricsConfidenceMetricsEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metrics for a single confidence threshold.
}
\concept{ClassificationEvaluationMetricsConfidenceMetricsEntry functions}
