% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageClassificationModelMetadata}
\alias{ImageClassificationModelMetadata}
\title{ImageClassificationModelMetadata Object}
\usage{
ImageClassificationModelMetadata(
  trainCost = NULL,
  nodeQps = NULL,
  baseModelId = NULL,
  modelType = NULL,
  stopReason = NULL,
  nodeCount = NULL,
  trainBudget = NULL
)
}
\arguments{
\item{trainCost}{Output only}

\item{nodeQps}{Output only}

\item{baseModelId}{Optional}

\item{modelType}{Optional}

\item{stopReason}{Output only}

\item{nodeCount}{Output only}

\item{trainBudget}{Required}
}
\value{
ImageClassificationModelMetadata object
}
\description{
ImageClassificationModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata for image classification.
}
\concept{ImageClassificationModelMetadata functions}
