% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ImageObjectDetectionModelMetadata}
\alias{ImageObjectDetectionModelMetadata}
\title{ImageObjectDetectionModelMetadata Object}
\usage{
ImageObjectDetectionModelMetadata(nodeCount = NULL, modelType = NULL,
  trainBudgetMilliNodeHours = NULL, stopReason = NULL,
  trainCostMilliNodeHours = NULL, nodeQps = NULL)
}
\arguments{
\item{nodeCount}{Output only}

\item{modelType}{Optional}

\item{trainBudgetMilliNodeHours}{The train budget of creating this model, expressed in milli node}

\item{stopReason}{Output only}

\item{trainCostMilliNodeHours}{Output only}

\item{nodeQps}{Output only}
}
\value{
ImageObjectDetectionModelMetadata object
}
\description{
ImageObjectDetectionModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to image object detection.
}
\concept{ImageObjectDetectionModelMetadata functions}
