% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ListDatasetsResponse}
\alias{ListDatasetsResponse}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
ListDatasetsResponse(datasets = NULL, nextPageToken = NULL)
}
\arguments{
\item{datasets}{The datasets read}

\item{nextPageToken}{A token to retrieve next page of results}
}
\value{
ListDatasetsResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2019-11-04 14:55:26
filename: /Users/justinmarciszewski/Dropbox/dev/R/autoGoogleAPI/googleautomlv1beta1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListDatasetsResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for AutoMl.ListDatasets.
}
\concept{ListDatasetsResponse functions}
