% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{ModelEvaluation}
\alias{ModelEvaluation}
\title{ModelEvaluation Object}
\usage{
ModelEvaluation(createTime = NULL, annotationSpecId = NULL,
  imageObjectDetectionEvaluationMetrics = NULL,
  textExtractionEvaluationMetrics = NULL,
  regressionEvaluationMetrics = NULL, name = NULL,
  evaluatedExampleCount = NULL,
  videoObjectTrackingEvaluationMetrics = NULL,
  textSentimentEvaluationMetrics = NULL,
  translationEvaluationMetrics = NULL,
  classificationEvaluationMetrics = NULL, displayName = NULL)
}
\arguments{
\item{createTime}{Output only}

\item{annotationSpecId}{Output only}

\item{imageObjectDetectionEvaluationMetrics}{Model evaluation metrics for image object detection}

\item{textExtractionEvaluationMetrics}{Evaluation metrics for text extraction models}

\item{regressionEvaluationMetrics}{Model evaluation metrics for Tables regression}

\item{name}{Output only}

\item{evaluatedExampleCount}{Output only}

\item{videoObjectTrackingEvaluationMetrics}{Model evaluation metrics for video object tracking}

\item{textSentimentEvaluationMetrics}{Evaluation metrics for text sentiment models}

\item{translationEvaluationMetrics}{Model evaluation metrics for translation}

\item{classificationEvaluationMetrics}{Model evaluation metrics for image, text, video and tables}

\item{displayName}{Output only}
}
\value{
ModelEvaluation object
}
\description{
ModelEvaluation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Evaluation results of a model.
}
\concept{ModelEvaluation functions}
