% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesModelMetadata}
\alias{TablesModelMetadata}
\title{TablesModelMetadata Object}
\usage{
TablesModelMetadata(optimizationObjective = NULL,
  optimizationObjectivePrecisionValue = NULL,
  tablesModelColumnInfo = NULL, inputFeatureColumnSpecs = NULL,
  trainBudgetMilliNodeHours = NULL, targetColumnSpec = NULL,
  optimizationObjectiveRecallValue = NULL, disableEarlyStopping = NULL,
  trainCostMilliNodeHours = NULL)
}
\arguments{
\item{optimizationObjective}{Objective function the model is optimizing towards}

\item{optimizationObjectivePrecisionValue}{Required when optimization_objective is 'MAXIMIZE_RECALL_AT_PRECISION'}

\item{tablesModelColumnInfo}{Output only}

\item{inputFeatureColumnSpecs}{Column specs of the dataset's primary table's columns, on which}

\item{trainBudgetMilliNodeHours}{Required}

\item{targetColumnSpec}{Column spec of the dataset's primary table's column the model is}

\item{optimizationObjectiveRecallValue}{Required when optimization_objective is 'MAXIMIZE_PRECISION_AT_RECALL'}

\item{disableEarlyStopping}{Use the entire training budget}

\item{trainCostMilliNodeHours}{Output only}
}
\value{
TablesModelMetadata object
}
\description{
TablesModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Tables.
}
\concept{TablesModelMetadata functions}
