% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesModelMetadata}
\alias{TablesModelMetadata}
\title{TablesModelMetadata Object}
\usage{
TablesModelMetadata(inputFeatureColumnSpecs = NULL,
  tablesModelColumnInfo = NULL, trainBudgetMilliNodeHours = NULL,
  targetColumnSpec = NULL, disableEarlyStopping = NULL,
  trainCostMilliNodeHours = NULL, optimizationObjective = NULL)
}
\arguments{
\item{inputFeatureColumnSpecs}{Column specs of the dataset's primary table's columns, on which}

\item{tablesModelColumnInfo}{Output only}

\item{trainBudgetMilliNodeHours}{Required}

\item{targetColumnSpec}{Column spec of the dataset's primary table's column the model is}

\item{disableEarlyStopping}{Use the entire training budget}

\item{trainCostMilliNodeHours}{Output only}

\item{optimizationObjective}{Objective function the model is optimizing towards}
}
\value{
TablesModelMetadata object
}
\description{
TablesModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Tables.
}
\concept{TablesModelMetadata functions}
