% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{TablesModelMetadata}
\alias{TablesModelMetadata}
\title{TablesModelMetadata Object}
\usage{
TablesModelMetadata(tablesModelColumnInfo = NULL,
  inputFeatureColumnSpecs = NULL, trainBudgetMilliNodeHours = NULL,
  targetColumnSpec = NULL, optimizationObjectiveRecallValue = NULL,
  trainCostMilliNodeHours = NULL, disableEarlyStopping = NULL,
  optimizationObjective = NULL,
  optimizationObjectivePrecisionValue = NULL)
}
\arguments{
\item{tablesModelColumnInfo}{Output only}

\item{inputFeatureColumnSpecs}{Column specs of the dataset's primary table's columns, on which}

\item{trainBudgetMilliNodeHours}{Required}

\item{targetColumnSpec}{Column spec of the dataset's primary table's column the model is}

\item{optimizationObjectiveRecallValue}{Required when optimization_objective is 'MAXIMIZE_PRECISION_AT_RECALL'}

\item{trainCostMilliNodeHours}{Output only}

\item{disableEarlyStopping}{Use the entire training budget}

\item{optimizationObjective}{Objective function the model is optimizing towards}

\item{optimizationObjectivePrecisionValue}{Required when optimization_objective is 'MAXIMIZE_RECALL_AT_PRECISION'}
}
\value{
TablesModelMetadata object
}
\description{
TablesModelMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata specific to AutoML Tables.
}
\concept{TablesModelMetadata functions}
