% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoObjectTrackingAnnotation}
\alias{VideoObjectTrackingAnnotation}
\title{VideoObjectTrackingAnnotation Object}
\usage{
VideoObjectTrackingAnnotation(instanceId = NULL, boundingBox = NULL,
  timeOffset = NULL, score = NULL)
}
\arguments{
\item{instanceId}{Optional}

\item{boundingBox}{Required}

\item{timeOffset}{Required}

\item{score}{Output only}
}
\value{
VideoObjectTrackingAnnotation object
}
\description{
VideoObjectTrackingAnnotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Annotation details for video object tracking.
}
\concept{VideoObjectTrackingAnnotation functions}
