% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl_objects.R
\name{VideoObjectTrackingDatasetMetadata}
\alias{VideoObjectTrackingDatasetMetadata}
\title{Cloud AutoML API Objects 
Train high-quality custom machine learning models with minimum effort and machine learning expertise.}
\usage{
VideoObjectTrackingDatasetMetadata()
}
\value{
VideoObjectTrackingDatasetMetadata object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2021-04-16 10:01:23
filename: /Users/justin/dev/R/autoGoogleAPI/googleautomlv1beta1.auto/R/automl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
VideoObjectTrackingDatasetMetadata Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dataset metadata specific to video object tracking.
}
\concept{VideoObjectTrackingDatasetMetadata functions}
